backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.25

# AI & Image Processing
openai==1.10.0
Pillow==10.2.0
imagehash==4.3.1
numpy==1.26.3

# Utilities
aiofiles==23.2.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# Redis for job queue
redis==5.0.1
rq==1.16.0

# EXIF data
piexif==1.1.3
exifread==3.0.0

# Progress tracking
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
